Move Logging api to a common place
