+++ examples/chef/efr32/include/AppConfig.h
@@ -18,6 +18,7 @@
  */
 
 #pragma once
+#include "silabs_utils.h"
 
 // ---- Lighting Example App Config ----
 
@@ -26,21 +27,3 @@
 // Time it takes in ms for the simulated actuator to move from one
 // state to another.
 #define ACTUATOR_MOVEMENT_PERIOS_MS 10
-
-// EFR Logging
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void silabsInitLog(void);
-
-void efr32Log(const char * aFormat, ...);
-#define SILABS_LOG(...) efr32Log(__VA_ARGS__);
-void appError(int err);
-
-#ifdef __cplusplus
-}
-
-#include <lib/core/CHIPError.h>
-void appError(CHIP_ERROR error);
-#endif
