+++ examples/lighting-app/silabs/efr32/include/AppConfig.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "efr32_utils.h"
+#include "silabs_utils.h"
 
 // ---- Lighting Example App Config ----
