+++ examples/platform/silabs/SiWx917/BUILD.gn
@@ -240,6 +240,7 @@ source_set("siwx917-common") {
   sources = [
     "${silabs_common_plat_dir}/LEDWidget.cpp",
     "${silabs_common_plat_dir}/heap_4_silabs.c",
+    "${silabs_common_plat_dir}/silabs_utils.cpp",
     "${silabs_common_plat_dir}/syscalls_stubs.cpp",
     "${wifi_sdk_dir}/dhcp_client.cpp",
     "${wifi_sdk_dir}/ethernetif.cpp",
@@ -249,7 +250,6 @@ source_set("siwx917-common") {
     "SiWx917/wfx_rsi_host.c",
     "init_ccpPlatform.cpp",
     "matter_config.cpp",
-    "siwx917_utils.cpp",
   ]
 
   if (use_base_app) {
