+++ examples/platform/silabs/efr32/BUILD.gn
@@ -282,8 +282,8 @@ source_set("efr32-common") {
   sources = [
     "${silabs_common_plat_dir}/LEDWidget.cpp",
     "${silabs_common_plat_dir}/heap_4_silabs.c",
+    "${silabs_common_plat_dir}/silabs_utils.cpp",
     "${silabs_common_plat_dir}/syscalls_stubs.cpp",
-    "efr32_utils.cpp",
     "init_efrPlatform.cpp",
     "matter_config.cpp",
   ]
